+++ package.json
@@ -18,44 +18,44 @@
 		"stylelint:fix": "stylelint \"**/*.{vue,css}\" --fix"
 	},
 	"dependencies": {
-		"axios": "^1.4.0",
-		"markdown-it": "^13.0.1",
+		"axios": "^1.6.0",
+		"markdown-it": "^13.0.2",
 		"material-design-icons-iconfont": "^6.7.0",
-		"pinia": "^2.1.6",
+		"pinia": "^2.1.7",
 		"query-string": "^8.1.0",
-		"ua-parser-js": "^1.0.35",
-		"vue": "^3.3.4",
-		"vue-router": "^4.2.4",
-		"vuetify": "^3.3.12",
+		"ua-parser-js": "^1.0.37",
+		"vue": "^3.3.7",
+		"vue-router": "^4.2.5",
+		"vuetify": "^3.3.23",
 		"workbox-window": "^7.0.0"
 	},
 	"devDependencies": {
-		"@types/markdown-it": "^13.0.1",
-		"@types/node": "^20.5.0",
-		"@types/ua-parser-js": "^0.7.36",
-		"@typescript-eslint/eslint-plugin": "^6.4.0",
-		"@typescript-eslint/parser": "^6.4.0",
+		"@types/markdown-it": "^13.0.5",
+		"@types/node": "^20.8.10",
+		"@types/ua-parser-js": "^0.7.38",
+		"@typescript-eslint/eslint-plugin": "^6.9.1",
+		"@typescript-eslint/parser": "^6.9.1",
 		"@vitejs/plugin-basic-ssl": "^1.0.1",
-		"@vitejs/plugin-vue": "^4.2.3",
-		"@vue/eslint-config-typescript": "^11.0.3",
-		"eslint": "^8.47.0",
+		"@vitejs/plugin-vue": "^4.4.0",
+		"@vue/eslint-config-typescript": "^12.0.0",
+		"eslint": "^8.52.0",
 		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-prettier": "^5.0.0",
-		"eslint-plugin-vue": "^9.17.0",
-		"postcss": "^8.4.27",
+		"eslint-plugin-prettier": "^5.0.1",
+		"eslint-plugin-vue": "^9.18.1",
+		"postcss": "^8.4.31",
 		"postcss-html": "^1.5.0",
-		"postcss-preset-env": "^9.1.1",
-		"prettier": "^3.0.2",
-		"stylelint": "^15.10.2",
+		"postcss-preset-env": "^9.3.0",
+		"prettier": "^3.0.3",
+		"stylelint": "^15.11.0",
 		"stylelint-config-recommended-vue": "^1.5.0",
 		"stylelint-config-standard": "^34.0.0",
-		"typescript": "5.1.6",
-		"vite": "^4.4.9",
+		"typescript": "5.2.2",
+		"vite": "^4.5.0",
 		"vite-plugin-eslint": "^1.8.1",
-		"vite-plugin-pwa": "^0.16.4",
-		"vite-plugin-stylelint": "^4.3.0",
+		"vite-plugin-pwa": "^0.16.6",
+		"vite-plugin-stylelint": "^5.2.0",
 		"vite-plugin-vuetify": "^1.0.2",
-		"vitest": "^0.34.1",
-		"vue-tsc": "^1.8.8"
+		"vitest": "^0.34.6",
+		"vue-tsc": "^1.8.22"
 	}
 }
