+++ src/App.vue
@@ -1,40 +1,13 @@
 <script setup lang="ts">
-import { computed, onMounted } from 'vue';
-import { IrdomNavbar, NavbarItem, IrdomToastList } from './components';
+import { onMounted } from 'vue';
+import { IrdomNavbar, IrdomToastList } from './components';
 import { useProfileStore, useTimetableStore } from './store';
 import { marketingApi } from './api/marketing';
+import { navbarItems } from '@/constants';
 
 const profileStore = useProfileStore();
 const { updateToken, updateTokenScopes, updateMarketingId } = profileStore;
 const { updateGroup } = useTimetableStore();
-
-const navbarItems = computed<NavbarItem[]>(() => {
-	const common = [
-		{
-			icon: 'calendar_month',
-			name: 'Расписание',
-			path: ['/timetable'],
-		},
-		{
-			icon: 'dashboard',
-			name: 'Сервисы',
-			path: ['/apps'],
-		},
-		// {
-		// 	icon: 'account_circle',
-		// 	name: 'Профиль',
-		// 	path: ['/profile', '/auth'],
-		// },
-	];
-	// if (profileStore.isAdmin) {
-	// 	common.push({
-	// 		icon: 'manage_accounts',
-	// 		name: 'Админка',
-	// 		path: ['/admin'],
-	// 	});
-	// }
-	return common;
-});
 
 onMounted(async () => {
 	updateToken();
@@ -49,6 +22,7 @@ onMounted(async () => {
 	}
 });
 </script>
+
 <template>
 	<v-app>
 		<RouterView />
