+++ src/constants/navbarItems.ts
@@ -1,0 +1,14 @@
+import { NavbarItem } from '@/components';
+
+export const navbarItems: NavbarItem[] = [
+	{
+		icon: 'calendar_month',
+		name: 'Расписание',
+		path: ['/timetable'],
+	},
+	{
+		icon: 'dashboard',
+		name: 'Сервисы',
+		path: ['/apps'],
+	},
+];
