+++ src/public/reset.html
@@ -11,7 +11,6 @@
 				caches
 					.keys()
 					.then(cacheNames => {
-						console.log(cacheNames);
 						cacheNames.forEach(cacheName => {
 							caches.delete(cacheName);
 						});
